refactor: move DetailedView to widgets and update import path
