+++ lib/app/widgets/detailedview.dart
